requirements.txt
aiohttp
httpx
python-dotenv
openai
groq
pydub